Frontend creation

+++ backend/controllers/product_controller.js
@@ -50,14 +50,18 @@ export const createProducts = async (req, res) => {
   }
 };
 
-
 export const deleteProducts = async (req, res) => {
   const { id } = req.params;
+  if (!id) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Product ID not found" });
+  }
 
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
